frontend/requirements.txt
# Core Streamlit
streamlit>=1.31.0

# HTTP and API Communication
requests>=2.27.0
aiohttp>=3.8.0
httpx>=0.24.1

# Environment and Configuration
python-dotenv==0.19.0
pydantic>=1.10.0,<2.0.0

# Data Processing and Visualization
numpy>=1.21.0
pandas>=1.3.0
Pillow>=10.4.0
matplotlib>=3.10.3
plotly>=5.0.0
seaborn>=0.11.0

# Async and Performance Optimization
asyncio-throttle>=1.0.0
cachetools>=5.0.0
aiocache>=0.12.0

# Data Validation and Serialization
protobuf>=3.20.0
jsonschema>=4.17.0

# Date and Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Security and Cryptography
cryptography>=41.0.0

# Logging and Monitoring
structlog>=23.1.0
rich>=13.3.0

# Development and Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0

# Additional Utilities
click>=8.1.0
tqdm>=4.65.0
